dbeaver/pro#4179 Refactor RPC bundle
- Delete module com.dbeaver.rpc.api (move to jkiss.utils)
- Rename module com.dbeaver.rpc -> com.dbeaver.ws.api
