--- modules/com.dbeaver.rpc.api/META-INF/MANIFEST.MF
@@ -1,11 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: DBeaver RPC API
-Bundle-Vendor: DBeaver Corp
-Bundle-SymbolicName: com.dbeaver.rpc.api
-Bundle-Version: 2.4.0.qualifier
-Bundle-Release-Date: 20240205
-Bundle-RequiredExecutionEnvironment: JavaSE-11
-Require-Bundle: org.jkiss.utils
-Export-Package: com.dbeaver.rpc.api
-Automatic-Module-Name: com.dbeaver.rpc.api
