--- modules/com.dbeaver.rpc/build.properties
@@ -1,3 +1,0 @@
-source.. = src/
-output.. = target/classes/
-bin.includes = META-INF/,.
