--- modules/com.dbeaver.rpc/src/com/dbeaver/rpc/ws/WsRuntimeException.java
@@ -1,27 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp
- *
- * All Rights Reserved.
- *
- * NOTICE:  All information contained herein is, and remains
- * the property of DBeaver Corp and its suppliers, if any.
- * The intellectual and technical concepts contained
- * herein are proprietary to DBeaver Corp and its suppliers
- * and may be covered by U.S. and Foreign Patents,
- * patents in process, and are protected by trade secret or copyright law.
- * Dissemination of this information or reproduction of this material
- * is strictly forbidden unless prior written permission is obtained
- * from DBeaver Corp.
- */
-package com.dbeaver.rpc.ws;
-
-public class WsRuntimeException extends RuntimeException {
-    public WsRuntimeException(String message) {
-        super(message);
-    }
-
-    public WsRuntimeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
