+++ modules/com.dbeaver.ws.api/build.properties
