+++ modules/pom.xml
@@ -19,8 +19,7 @@
     <modules>
         <module>org.jkiss.utils</module>
         <module>com.dbeaver.jdbc.api</module>
-        <module>com.dbeaver.rpc.api</module>
-        <module>com.dbeaver.rpc</module>
+        <module>com.dbeaver.ws.api</module>
     </modules>
 
     <build>
